requirements.txt
streamlit
feedparser
trafilatura
google-generativeai
beautifulsoup4
lxml_html_clean